split up autocrypt operations class into package
